Added log in functionality with cookies
used the same form data handling as we learned in class.

then we pull local storage data and check that the email exists in the local data, if the email exists we check for matching passwords. If everything checks out we set user cookie data so that profile information etc. can be displayed

right now I did a lot of the cookie stuff manually without functions just to get some things working.

The page currently does not change locations after log in so it's likely working just not relocating yet. This can be added.

If email is not found in local data OR if passwords do not match there is an alert.