Updated signup page to load local storage user data
create users with useState

useEffect #1, on load we collect local storage userdata and and then save it to our users state.

useEffect #2 fires when users state is updated. It checks if user data already exists, if not this does nothing extra.
If users data exists (meaning nothing went wrong with data collection or addition) then it sets the local storage with the new data including the new user

addUser function now checks the localstorage for already existing email, if not it adds the user to the local storage.

No page change on sign up right now.