added new navbar component for student users
This is current working when logged in as student. I think we will have to create some dummy data for pre-defined admins since there is no way to create an admin account.

Also I think I will change the way this works and put all the checks within the nav bar component then return the correct links for the type of user.

Currently the header component checks type of user then returns the correct burger menu with navbar component